+++ MyUECppProject/Source/MyUECppProject/Component/CircleSplineComponent.h
@@ -22,7 +22,7 @@ protected:
 
 public:
 	// Called every frame
-	virtual void TickComponent(double DeltaTime, ELevelTick TickType, FActorComponentTickFunction* ThisTickFunction) override;
+	virtual void TickComponent(float DeltaTime, ELevelTick TickType, FActorComponentTickFunction* ThisTickFunction) override;
 
 #if WITH_EDITOR
 	virtual void PostEditChangeProperty(FPropertyChangedEvent& PropertyChangedEvent) override;
